style: move `utils` to their own import group
This patch updates my ESLint configuration to move `utils` into their
own import group. This patch also moves the `utils.ts` file to
`utils/general` so that it can also be included in that group.

Closes: NC-712
