+++ app/routes/_header.shows.$seasonYear.$seasonName.($sex).$brandSlug/looks.tsx
@@ -18,9 +18,9 @@ import {
 } from 'components/ui/command'
 import { Popover, PopoverContent, PopoverTrigger } from 'components/ui/popover'
 
-import { useOptionalUser, useRedirectTo, uniq } from 'utils'
 import { cn } from 'utils/cn'
 import { commandScore } from 'utils/command-score'
+import { useOptionalUser, useRedirectTo, uniq } from 'utils/general'
 
 import { type loader } from './route'
