+++ app/routes/api.logout.ts
@@ -1,7 +1,8 @@
 import { type ActionArgs } from '@vercel/remix'
 
+import { safeRedirect } from 'utils/general'
+
 import { logout } from 'session.server'
-import { safeRedirect } from 'utils'
 
 export async function action({ request }: ActionArgs) {
   const formData = await request.formData()
