+++ app/utils/general.ts
